nannou_core
Module geom

Types, functions and other items related to geometry. This module is the source of all graphics and lazer primitives and aids work in 2D and 3D space.

Each module provides a set of general tools for working with the named geometry including:

A typed, object representation.

Functions for producing vertices, triangles and triangulation indices.

Functions for checking whether or not the geometry contains a point.

Functions for determining the bounding rectangle or cuboid.

A function for finding the centroid.

Re-exports

pub use self::cuboid::Cuboid;

pub use self::ellipse::Ellipse;

pub use self::point::pt2;

pub use self::point::pt3;

pub use self::point::pt4;

pub use self::point::Point2;

pub use self::point::Point3;

pub use self::point::Point4;

pub use self::polygon::Polygon;

pub use self::quad::Quad;

pub use self::range::Align;

pub use self::range::Edge;

pub use self::range::Range;

pub use self::rect::Corner;

pub use self::rect::Padding;

pub use self::rect::Rect;

pub use self::scalar::Scalar;

pub use self::tri::Tri;

pub use self::vector::Vector2;

Deprecated

pub use self::vector::Vector3;

Deprecated

pub use self::vector::Vector4;

Deprecated

pub use self::vertex::Vertex;

pub use self::vertex::Vertex2d;

pub use self::vertex::Vertex3d;

Modules

cuboid

Items related to cube geometry.

ellipse

point

Implementation of the Point types.

polygon

quad

range

A type for working with one-dimensional ranges.

rect

scalar

tri

vector

vertex

Structs

DVec2

A 2-dimensional vector.

DVec3

A 3-dimensional vector.

DVec4

A 4-dimensional vector.

IVec2

A 2-dimensional vector.

IVec3

A 3-dimensional vector.

IVec4

A 4-dimensional vector.

Vec2

A 2-dimensional vector.

Vec3

A 3-dimensional vector without SIMD support.

Vec4

A 4-dimensional vector.

Functions

bounding_cuboid

The Cuboid that bounds the given sequence of vertices.

bounding_rect

The Rect that bounds the given sequence of vertices.

centroid

The centroid (average position) of all vertices in the given iterator.

dvec2

Creates a 2-dimensional vector.

dvec3

Creates a 3-dimensional vector.

dvec4

Creates a 4-dimensional vector.

ivec2

Creates a 2-dimensional vector.

ivec3

Creates a 3-dimensional vector.

ivec4

Creates a 4-dimensional vector.

vec2

Creates a 2-dimensional vector.

vec3

Creates a 3-dimensional vector.

vec4

Creates a 4-dimensional vector.