nannou_core::geom
Module cuboid

Items related to cube geometry.

The main type is the Cuboid type.

Structs

Corners

An iterator yielding each corner of a cuboid in the following order.

Cuboid

A light-weight Cuboid type with many helper and utility methods.

FaceQuads

An iterator yielding each face of a cuboid as a quad.

Faces

An iterator yielding the faces of a cuboid as per their ordering.

SubdivisionRanges

The three ranges that make up the 8 subdivisions of a cuboid.

Subdivisions

Yields even subdivisions of a Cuboid.

Triangles

An iterator yielding all triangles for all faces.

Enums

Face

Each of the faces of a cuboid.

Constants

NUM_CORNERS

The number of corners on a Cuboid.

NUM_FACES

The number of faces on a Cuboid.

NUM_SUBDIVISIONS

The number of subdivisions for a Cuboid.

NUM_TRIANGLES

The number of triangles used to triangulate a cuboid.

Type Aliases

FaceQuad

A quad representing a single face of a cuboid.