nannou_core::math
Trait Mat4LookTo

pub trait Mat4LookTo {
    // Provided methods
    fn look_to_rh(eye: Vec3, dir: Vec3, up: Vec3) -> Mat4 { ... }
    fn look_to_lh(eye: Vec3, dir: Vec3, up: Vec3) -> Mat4 { ... }
}

Create a transformation matrix that will cause a vector to point at dir using up for orientation.

Provided Methods

fn look_to_rh(eye: Vec3, dir: Vec3, up: Vec3) -> Mat4

fn look_to_lh(eye: Vec3, dir: Vec3, up: Vec3) -> Mat4

Object Safety

This trait is not object safe.